Use array storage in Vec3
I realized that since we know the runtime number of elements, there's no
need to avoid using arrays (which was what I did in the Ray Tracer
Challenge implementation). This is great because it's easier to use the
book's implementations as-is. We start by updating AABB to use the
book's implementation.